+++ src/org/jruby/ext/posix/LinuxFileStat.java
@@ -50,7 +50,7 @@ public class LinuxFileStat extends BaseNativeFileStat {
         return st_gid;
     }
 
-    public int ino() {
+    public long ino() {
         return st_ino;
     }
