+++ src/org/jruby/ext/posix/LinuxHeapFileStat.java
@@ -62,7 +62,7 @@ public class LinuxHeapFileStat extends BaseHeapFileStat {
         return st_gid.get();
     }
 
-    public int ino() {
+    public long ino() {
         return st_ino.get();
     }
