+++ src/org/jruby/ext/posix/WindowsFileStat.java
@@ -60,7 +60,7 @@ public class WindowsFileStat extends BaseNativeFileStat {
         return st_gid;
     }
 
-    public int ino() {
+    public long ino() {
 	return st_ino;
     }
